Removed some code duplication and more
Moved the call to TMI into its own promise so it can be sharded between the two api calls

added the capture of "no mods" in a channel.

added the channel owner to the list of mods if they are not a mod already.

sorted the mod list.

drank a beer.